Remove sweetAlert() usage to dramatically reduce component overhead
