+++ lang/en.json
@@ -721,13 +721,7 @@
         },
         "add_to_cart": "Add to Cart",
         "adding_to_cart": "Adding to cart…",
-        "buy_now": {
-            "pdp-button-text": "Buy Now",
-            "modal-title": "Are you ready to checkout?",
-            "modal-text": "You're about to be taken to the checkout page",
-            "modal-cancel-text": "No, continue shopping",
-            "modal-confirm-text": "Yes, let's do this"
-        },
+        "buy_now_button": "Buy Now",
         "options": "Options",
         "none": "None",
         "sku": "SKU:",
